ChatBotHub.AI/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

chromadb==0.5.23
fastapi==0.115.6
pydantic==2.10.4
starlette==0.41.3

llama-index-core~=0.12.8
python-dotenv~=1.1.0
sympy~=1.14.0